add TDD Domain.Command

+++ src/Pedidos.Domain.Command/Commands/ItemPedido/ItemPedidoCommand.cs
@@ -3,7 +3,7 @@
     public class ItemPedidoCommand
     {
         public string Descricao { get; set; }
-        public double PrecoUnitario { get; set; }
+        public decimal PrecoUnitario { get; set; }
         public int Quantidade { get; set; }
     }
 }
